quick fix to Migration landing page (#1503)
## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
Migration landing page is currently out of step with other pages on
Academy; specifically, it's showing descriptions of each doc listed,
making the page cluttered and difficult to read.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
Removes the `description` value from the _index.md file's front matter.

### Why are we making this change?
<!-- What larger problem does this PR address? -->

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
Should look okay in the deploy preview.

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
LMK if this is okay. I don't think we've reached a critical mass of docs
in this bucket to make a fancier landing page like the one for Images.

Signed-off-by: Mark Drake <mark@chainguard.dev>