IAM docs update, focus on roles and role-bindings (#1493)
## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
The main goal of this PR is to add a doc outlining what roles and
role-bindings are within the context of Chainguard. As part of this, I
pulled the "role bindings" section (actually focused on roles, not role
bindings) from the IAM overview and reworked it to be more correct, and
also expanded on it to flesh out the overview.

I also added a new section about identities to the IAM overview. I also
thought it made sense to remove "enforce" from the IAM model overview
doc's URL and add a redirect

Lastly, I did a code search to change any instance of "rolebinding" or
"role binding" to "role-binding".

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
Resolves https://github.com/chainguard-dev/internal/issues/3548

### Why are we making this change?
<!-- What larger problem does this PR address? -->
Improving our IAM documentation

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
It should read well, changes should make sense, and the commands in the
new doc should work!

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
Copyedit and tech test, please!

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>