Getting Started with the nginx Chainguard Image Tutorial (#1585)
## Type of change
Addition to documentation - adding new getting started tutorial

### What should this PR do?
Adds a new getting started tutorial to the Academy for the nginx
Chainguard Image
https://github.com/chainguard-dev/internal/issues/3904

### Why are we making this change?
Expands Academy content that is helpful for new users, especially if a
user may be interested in using an image for their nginx application(s)

### What are the acceptance criteria? // How should this PR be tested?
1. Please try to run this demonstration locally. I have had success with
it multiple times on my MacOS laptop but I have not had an external test
yet
2. Please read through the tutorial for any mistakes or for any
suggestions for wording changes. I aimed to follow the style of the
other "getting started" guides, but there has not been a formal review
yet here

> Note: I did not create a demo with a multi-stage build. If we want to
add something more complex, I am happy to try and figure that out. I
wasn't sure how long we wanted this to be. It already seemed lengthy to
me.

3. There is a pull request here
https://github.com/chainguard-dev/edu-images-demos/pull/17 for the demo
tutorial components which should be approved & merged before this
article goes live since there is a reference to the repo link in the
guide
4. Once the demo is tested, the other request merged, and any
suggestions applied, it should be ready to go live

---------

After quite a few edits & testing a few extra notes were added to address potential permission errors that may occur. This should be in a good place to release.

Signed-off-by: mcaveety <michelle.mcaveety@chainguard.dev>