adding chainctl install instructions for Windows (#1635)
## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
Minor change to update the `chainctl` installation doc to include
instructions for installing on Windows. The method outlined here uses
`curl` to install it in PowerShell

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
Resolves https://github.com/chainguard-dev/edu/issues/1618 

### Why are we making this change?
<!-- What larger problem does this PR address? -->
To support our wonderful Windows users

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
Flagging @johnfosborneiii as he's the source of this request.

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
I set up a Windows 11 VM to confirm that this does indeed work, though I
am a longtime Windows avoider so any feedback is welcome.

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>