minor changes for grammar (#1722)
## Type of change

I made small corrections on two pages. This is my first PR to this repo
(Yay!) and I'm making a super-small one to test whether I have my access
set up properly.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->

n/a

### Why are we making this change?

I don't yet know which style guide we use, if we use one, but most use
"log in/log off" as the imperative verb form and "login/logoff" as
nouns. The Microsoft style guide wants everyone to shift to "sign in"
and "sign off" but gives an exception for matching the text used to the
UI.

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->

Acceptance criteria is whether or not my change fits the style guide in
use.

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->

Confirm this matches our style guide.