Remove terminal from policy-controller pages, it isn't working (#1731)
## Type of change
content/interactivity

### What should this PR do?
This disables the terminal on policy-controller pages. 

### Why are we making this change?
For some reason the kubectl alias isn't working. It is also running an
out of date k3s.

### What are the acceptance criteria? 
Content still reads well without the terminal.

### How should this PR be tested?
This won't show on netlify, but on deploy the terminal won't load on
policy-controller related pages.

Signed-off-by: Jamon <jamon@chainguard.dev>