landing page for working w/ imgs bucket (#1743)
## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
I thought we reached enough docs in the Working With Images bucket that
we could have a nice landing page for it.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
Resolves #1742 

### Why are we making this change?
<!-- What larger problem does this PR address? -->
Just a visual improvement over the current page.

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
should look good and the buttons/everything should work

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
Just give it a visual review. I also removed what seemed to be an
extraneous block of text from the `Getting Started Guides` landing page
(which I used as the basis for this page) but lmk if that should go back

Landing page can be found here:
https://deploy-preview-1743--ornate-narwhal-088216.netlify.app/chainguard/chainguard-images/working-with-images/

PS - I chose the cards I did because they seemed 1. relevant/useful and
2. they had short titles (or linktitles) that I thought would fit and
also be descriptive

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>