Updating imgs directory walkthrough with recent console changes (#1752)
## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
This updates the walkthrough for the Images Directory to include some
newly-added details, including the Entitlements column, as well as the
Specifications and Advisories tabs.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
Resolves https://github.com/chainguard-dev/internal/issues/3839

### Why are we making this change?
<!-- What larger problem does this PR address? -->
This was originally supposed to highlight some recent changes to the
Console but I ended up redoing all the screenshots as otherwise it'd be
a mix of the old/new branding.

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->

Give the changes a read and let me know what you think! Also please pay
close attention to the screenshots.

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>
Signed-off-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>