Add technical article on security advisories (#1747)
There was a bit of back-and-forth regarding whether or not we should
include the technical guide for managing security advisories on the
Academy, see #1714 for the context.

After discussing with @sheesh and others, I believe this article should
live on the Academy. The security advisories bucket is a logical place
for it to go. This article will benefit Chainguard employees and placing
it here will make it more accessible and intuitive to find.

There was concern that this article was only suited for internal use,
however,
a) external users have contributed in the past, although rarely
b) if a user wanted to contribute to the advisories repo, they need to
do further reading as stated by the article
c) any submissions to the repository are audited by an employee before
they are merged

I don't think this woud necessarily *encourage* external users to
contribute or cause noise. If anything, external contributions still
benefit anyone using images.

To review this article, please test the site locally on your machine and
follow the guide. It is suggested that you install wolfictl using the
instructions in this tutorial and try out the commands explored in this
guide.

@SharpRake for review :D sorry, I know I've sent you a lot in a short
amount of time! Please take your time 🙏

---------

Signed-off-by: mcaveety <michelle.mcaveety@chainguard.dev>
Signed-off-by: Michelle McAveety <135155059+mcaveety@users.noreply.github.com>
Co-authored-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>