Minor improvements - adding video embed and switching Node --> Go (#1764)
This PR achieves two things:
1. Embeds the Getting Started with the nginx Chainguard Image video into
its associated guide for more visibility. (note: no transcript provided
at this time since it walks through the written tutorial content, though
I could write something up if it's desired)
https://deploy-preview-1764--ornate-narwhal-088216.netlify.app/chainguard/chainguard-images/getting-started/nginx/

2. Fixes a reference to Node that should say Go. See the diff in the
file for more context. It seems like the wrong image was referenced
(likely because Node showed up earlier in advisories database) but it
appears that the Go image is what is used for the hands-on portion.
https://deploy-preview-1764--ornate-narwhal-088216.netlify.app/chainguard/chainguard-images/working-with-images/security-advisories/how-to-use/#comparing-images

To test this PR, it would be great if someone can double check that
these two changes look correct (especially the video) on the deploy
preview.

---------

Signed-off-by: mcaveety <michelle.mcaveety@chainguard.dev>
Signed-off-by: Michelle McAveety <135155059+mcaveety@users.noreply.github.com>
Co-authored-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>