Remove OOD info on enterprise. (#1737)
Deleted section on enterprise usage.

This section did explain how to use the API with a long-lived token,
which is still potentially useful, but it would need updated and
checked. The remaining instructions are preferable as they don't involve
long-lived tokens. Given the instructions have been broken for a
long-time and no-one complained, I don't think anyone will miss it.

Signed-off-by: Adrian Mouat <adrian@chainguard.dev>