Getting Started with the C/C++ Chainguard Image  (#1723)
## Type of change
Adding a new Getting Started guide to the Academy for C/C++ images!! 

### What should this PR do?
Resolves [#3888](https://github.com/chainguard-dev/internal/issues/3888)

### Why are we making this change?
Recently we published a [cheatsheet on the different uses of images for
various compiled
programs](https://edu.chainguard.dev/chainguard/chainguard-images/working-with-images/compiled-programs/)
so this document provides an example of how we can get started using
them!

### What are the acceptance criteria? 
The guide should be functional and it should follow a similar format to
the other guides we currently have. It should walk through using
Chainguard Images to form a single and multi-layer docker build

### How should this PR be tested?
Please go through the guide as a learner and test the commands. Please
also look for places where improvements can be made to the clarity of
the writing.

Link to the guide in the deploy preview:
https://deploy-preview-1723--ornate-narwhal-088216.netlify.app/chainguard/chainguard-images/getting-started/c/

---------

Signed-off-by: mcaveety <michelle.mcaveety@chainguard.dev>
Signed-off-by: Michelle McAveety <135155059+mcaveety@users.noreply.github.com>
Co-authored-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>