Tier changes, part 2/3 (#1779)
## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
This PR implements a number of changes outlined in my Tier audit:
https://docs.google.com/document/d/1Go2vRf40PlVWka41dG_jmNZY_xkhhBbclf28duWwZ1E/edit?usp=sharing

Specifically, this PR updates any resources that use an affected image
in an example to using one that will remain avaiable after the change.
It also fixes the vulnerability comparisions template to point readers
to the image's readme in the Image's Directory.

Part 1/3 is Erika's initial work to improve/clarify language around our
pricing tiers
Part 3/3 will be implementing a short code note and adding it to
resources that will not be updated and instead continue to use images.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
Resolves https://github.com/chainguard-dev/internal/issues/4126

### Why are we making this change?
<!-- What larger problem does this PR address? -->
Clarity around products and pricing

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
Changes should make sense. They're generally very small / trivial to the
bulk of the respective guide.

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
Give changes a close look please!

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>