Artifactory pull thru guide update (#1801)
## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
This PR adds a new section to the Artifactory Pull through doc that
highlights how to use one of Artifactory's virtual repositories as a
pull through cache for Chainguard Images.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
Resolves #1789 

### Why are we making this change?
<!-- What larger problem does this PR address? -->
Sales folks let us know customers were hitting an issue with this doc.

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
Testing this might be slightly annoying, as you'll need an Artifactory
account and project. I just set up a free trial, which worked well
enough for me.

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
The new section is somewhat built off of the previous sections in the
guide - it may be easier to run through the whole thing just to test the
new section, but your move.

You can find the preview here:
https://deploy-preview-1801--ornate-narwhal-088216.netlify.app/chainguard/chainguard-registry/pull-through-guides/artifactory-pull-through/

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>