add link to FIPS page to all mentions (#1862)
## Type of change

### What should this PR do?
This adds a link to our wonderful FIPS page from all other pages that
refer to FIPS in some way.
This is related to
https://github.com/chainguard-dev/internal/issues/4268

### Why are we making this change?
To lead people to the page that tells them all about FIPS so they are
less likely to be confused or misconfigure stuff.

### What are the acceptance criteria? 
I won't be mad if you check the links.

### How should this PR be tested?
Did I insert any new typos? Are my links accurate? Then, let's go!

---------

Signed-off-by: Matthew Helmke <matthew.helmke@chainguard.dev>
Co-authored-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>