initial fix for FIPS page (#1872)
## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
Documentation update/fix

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
Edits/replaces the content on the FIPS page

### Why are we making this change?
<!-- What larger problem does this PR address? -->
As time has passed, much of the information became outdated. This is a
stage 1 fix to bring the page into parity with
https://www.chainguard.dev/legal/fips-commitment so that we aren't
sending mixed messages. The plan is to continue to evolve this page as
time permits.

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
Is the information accurate? I'm working with @xnox on this, so when
he's happy we will merge.

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
n/a

---------

Signed-off-by: Matthew Helmke <matthew.helmke@chainguard.dev>
Co-authored-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>