Fixing invalid timestamps identified by maintenance tool (#1887)
As part of the work for bootstrapping our maintenance sprint, a series
of articles were identified with invalid timestamps. This PR fixes this
so we can get reliable information about last updated dates.

In addition to the timestamp changes, there are automatic changes made
by my editor to remove empty spaces from the files.