Maintenance tools (#1888)
This PR includes a custom tool under `autodocs/` to help with
maintenance and content audits. This is discussed in more detail in
internal issue
[4360](https://github.com/chainguard-dev/internal/issues/4360).

This is a CLI tool written in PHP. It can be expanded with more
utilities as we need it.

Having the tool alongside the content is a lot easier to manage, so
instead of creating a separate repository as I did in the past with
Autodocs, I reused the existing audodocs folder to set up this tool.