Maintenance - updating What is OpenVEX doc (#1925)
## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
Part of the maintenance sprint, this PR updates the `What is OpenVEX?`
doc.

I tested the few commands in the doc and validated the info. Most of my
effort was simplifying the language and breaking up paragraphs to make
them smaller and more digestible.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
resolves #1893 

### Why are we making this change?
<!-- What larger problem does this PR address? -->
Maintenance sprint!

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
Should meet expectations for maintenance work:
https://docs.google.com/document/d/1Cy4WWUHu_WE365tNqBiYBFQ8Y-OJtrdHUf2kBftq85c/edit?usp=sharing

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
Not much to tech test, but feel free if you'd like!

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>