Update fips-images.md lastmod (#1927)
Updating `lastmod` date as it was just recently updated and it's
included in our maintenance batch.

## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
resolves https://github.com/chainguard-dev/edu/issues/1890

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
All this PR does is change the `lastmod` date so we can say the
maintenance effort is completed.

### Why are we making this change?
<!-- What larger problem does this PR address? -->
Aaditya (with the support of Dimitri and Matthew H) updated this doc,
but the `lastmod` wasn't updated. This PR fixes that.

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
I have not performed any real review of the content, as that was
completed by Matthew, Dimitri, and Adrian I believe
(https://github.com/chainguard-dev/edu/pull/1912). All I'm doing is
updating the `lastmod` field to signify that.

Signed-off-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>