Adding guide on strats + tooling for updating images (#1938)
## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
This PR adds a modified blog post (originally by Adrian) to Academy on
strategies and tools for keeping container images up to date. This is
part of an effort to add more resources on how to handle EOL software to
Academy

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
This doesn't resolve any issues on its own.

### Why are we making this change?
<!-- What larger problem does this PR address? -->
Refer to [this PR for more
info](https://github.com/chainguard-dev/internal/issues/4355)

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
The content is ready for review though it may end up moving to a
slightly different location, depending on how the convo around
organization goes in our team meeting tomorrow

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
Give a close review for grammar, style, spelling, etc. This is a
simplified and more generalized version of an existing blog post
([reference](https://www.chainguard.dev/unchained/stay-secure-strategies-and-tooling-for-updating-container-images))
but some things may have been lost in translation.

preview:
https://deploy-preview-1938--ornate-narwhal-088216.netlify.app/chainguard/chainguard-images/recommended-practices/updating-images/strategies-tools-updating-images/

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>