Updated how to use securitiy advisories (#1948)
## Description 

Updates [How to use Chainguard Security Advisories and the Diff
API](https://edu.chainguard.dev/chainguard/chainguard-images/videos/security_advisories/)
(video page) and [How to Use Chainguard Security
Advisories](https://edu.chainguard.dev/chainguard/chainguard-images/working-with-images/security-advisories/how-to-use/)
as part of the maintenance sprint.

## Details

Since the assigned page was a video page, I updated the linked article
and then added a note to the video page, moving up the note and link to
the top.

On the article page, I simplified the intro and switched the examples to
use Grype, in my testing Docker Scout was acting a bit hinky on the old
tags.

## Notes

Resolves https://github.com/chainguard-dev/edu/issues/1901

Signed-off-by: Patrick Smyth <patrick.smyth@chainguard.dev>