Updating videos - maintenance sprint #1902 #1903 #1904 (#1963)
This PR is part of the Maintenance Sprint - #1902 #1903 #1904

I checked the 3 videos and they looked still valid and relevant, I
didn't see any major change that required a disclaimer or notice added
to the page. These are also not very old, they are from the beginning of
this year.

To make a little improvement I added a "Relevant Links" section at the
end of the article with links to our own resources that could compliment
the material. Since these weren't big changes, I grouped the 3 issues in
a single PR.