Fixing 404 links, 1-6-25 (#1991)
## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
This PR fixes some links that the checker has found to be 404ing. It
adds a few links to the ignore file as well.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
resolves https://github.com/chainguard-dev/edu/issues/1990

### Why are we making this change?
<!-- What larger problem does this PR address? -->
404s are bad!

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
Just want to note I'm still feeling out the link checker. I'm bad at
regexp but it would probably be cleaner to add the `0.0.0.0` urls there
once I'm able to figure that out.

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->

Signed-off-by: Mark Drake <mark@chainguard.dev>