Updating CVE Visualizations doc (#2054)
[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
This PR updates the CVE Visualizations doc to include the new features
that have been added since the Beta launch. It also reworks some of the
language and fleshes out a few explanations.



### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
Note that I'd like to redo some screenshots next week after the formal
launch so as not to include `BETA` in the screenshots, so unless there's
a good reason otherwise I'd like this not to resolve the issue I have
for this.

I'm also planning to update the images directory doc screenshots to show
the Reports tab once the beta flag is gone.

### Why are we making this change?
<!-- What larger problem does this PR address? -->
Feature is coming out of beta, let's flesh out the docs!

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
Changes and screenshots should look good.

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
No real tech test. If you'd like to explore the feature/check what i'm
talking about you can access it in the following locations:

Directory preprod:
https://images.preprod.chainguard.app/?feature.compareDirectory=true
Console preprod:

https://preprod.chainguard.app/?feature.forceData=true&feature.resolvedChart=true

Preview link for updates:
https://deploy-preview-2054--ornate-narwhal-088216.netlify.app/chainguard/chainguard-images/features/cve_visualizations/

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>
Signed-off-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>