Adding dfc EA user guide to Academy (#2126)
This PR adds the [Dockerfile Conversion
Tool](https://github.com/chainguard-dev/dfc) EA user guide to Academy,
under the "Migration" menu, with the path
`/chainguard/migration/dockerfile-conversion/` .

In the moment, we only have a EA user guide, so this is a single
markdown file. We can expand this later on as a subfolder with multiple
docs.

Preview:
https://deploy-preview-2126--ornate-narwhal-088216.netlify.app/chainguard/migration/dockerfile-conversion/

---------

Signed-off-by: Erika Heidi <erikaheidi@users.noreply.github.com>
Co-authored-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>