Footer update (#2147)
[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
This updates the footer to be consistent across web properties. 

### What should this PR do?
Update the footer. 

### Why are we making this change?
Consistency. 

### What are the acceptance criteria? 
Should match the footer on www. Links should all work and have utms
added.

### How should this PR be tested?
Please verify the links are pointing to the right places and that there
are no typos or other errors. Also a quick scan of the toml would be
appreciated.

Signed-off-by: ltagliaferri <lisa.tagliaferri@gmail.com>