Improve docs for Java repo manager config
## Type of change

### What should this PR do?

Improve docs for Java repo manager config as a result of further
reviewing and testing. Changes create more consistency and expand on the
necessary steps to get the build tool configuration set up. All that is
now also fully tested. The only issue is that Cloudsmith docs are
correct now .. but its not working. I am investigating with the
engineering team.

Part of https://github.com/chainguard-dev/internal/issues/4803

### Why are we making this change?

better is better

### What are the acceptance criteria? 

Passing review

### How should this PR be tested?

Not really necessary but if desired reviewer and follow setup steps with
desired repo manager.