add callout to privileges necessary in the Azure console (#2169)
- [x] Check if this is a typo or other quick fix and ignore the rest :)

Adding users to an Entra app requires admin permissions in Azure.
Without them, you'll be able to do every step on this page, but you will
not be able to add any users to the new application. This PR calls out
to setup the integration as an Admin user in Azure